Refactor calendar nav for composability
