+++ src/components/Calendar/CalendarPageProvider.vue
@@ -1,0 +1,14 @@
+<template>
+  <slot />
+</template>
+
+<script setup lang="ts">
+import { Page } from '../../utils/page';
+import { createPage } from '../../use/page';
+
+const props = defineProps<{
+  page: Page;
+}>();
+
+createPage(props.page);
+</script>
